Remove implicit chaining, as it has problems with ambiguity if a method can either return None or a value.
When it returned None, implicit chaining would kick in and the code would suddenly do something different, which is hard to debug. Instead the new symbol .__ is added to still allow chaining of None returning methods, but in an explicit way.

So _([1,2,3]).sort(reverse=True).__[0] == 3 is the explicit way to do this now.